+++ happy_seed.rb
@@ -3,35 +3,6 @@ require 'bundler'
 puts "Setting up basic template"
 puts
 
-gsub_file 'Gemfile', /.*sqlite3.*/, ""
-gem 'haml-rails'
-
-gem_group :development, :test do
-  gem "sqlite3"
-  gem "rspec"
-  gem "rspec-rails"
-  gem "factory_girl_rails"
-  gem "capybara"
-  gem "cucumber-rails", :require => false
-  gem "guard-rspec"
-  gem "guard-cucumber"
-  gem "database_cleaner"
-  gem "spring-commands-rspec"
-  gem 'spring-commands-cucumber'
-  gem "quiet_assets"
-  gem "launchy"
-  gem "vcr"
-  gem "faker"
-end
-
-gem_group :test do
-  gem "webmock"
-end
-
-gem_group :production do
-  gem 'pg'
-end
-
 gem_group :development do
   if ENV['SEED_DEVELOPMENT']
     gem 'happy_seed', :path => ENV['SEED_DEVELOPMENT'] # File.dirname(__FILE__)
@@ -40,101 +11,52 @@ gem_group :development do
   end
 end
 
-packages = []
+@packages = [ 'base' ]
+
+def run_graph graph
+  graph.each do |node|
+    if yes?( "Install #{node[:name]}, #{node[:desc]}?" )
+      generate "happy_seed:#{node[:name]}"
+      @packages << node[:name]
+
+      run_graph node[:subtree] if node[:subtree]
+    end
+  end
+end
 
 Bundler.with_clean_env do
   run "bundle install --without production"
 
-  gsub_file "app/assets/javascripts/application.js", /= require turbolinks/, "require turbolinks"
-
-  # Install rspec
-  generate "rspec:install"
-  gsub_file ".rspec", "--warnings\n", ""
-  append_to_file ".rspec", "--format documentation\n"
-
-  # Install cucumber
-  generate "cucumber:install"
-
-  append_to_file "features/support/env.rb", "\nWorld(FactoryGirl::Syntax::Methods)\n"
-
-  # Install Guard
-  run "guard init"
-
-  # Use the spring version and also run everything on startup
-  gsub_file "Guardfile", 'cmd: "bundle exec rspec"', 'cmd: "bin/rspec", all_on_start: true'
-  gsub_file "Guardfile", 'guard "cucumber"', 'guard "cucumber", cli: "--color --strict"'
-
   # Run the base generator
   generate "happy_seed:base"
 
-  all_in = false
-  all_in = true if yes? "Would you like to install everything?"
+  puts "Base generator installed."
 
-  if all_in || yes?( "Would you like to install jazz hands?" )
-    generate "happy_seed:jazz_hands"
-    packages << "jazz_hands"
-  end
+  puts File.read( File.expand_path( "happy_seed.txt", File.dirname( __FILE__ ) ) )
 
-  if all_in || yes?( "Would you like to install bootstrap?" )
-    generate "happy_seed:bootstrap"
-    packages << "bootstrap"
+  GRAPH = [
+    {name: 'splash', desc: 'Basic splash page' },
+    {name: 'devise', desc: 'User profiles', subtree: [
+      {name: 'devise_invitable', desc: 'Invitable users'},
+      {name: 'facebook', desc: 'OAuth: Connect with facebook' },
+      {name: 'github', desc: 'OAuth: Connect with github' },
+      {name: 'googleauth', desc: 'OAuth: Connect wuth google' },
+      {name: 'instagram', desc: 'OAuth: Instagram' },
+      {name: 'twitter', desc: 'OAuth: twitter' },
+      ]},
+    {name: 'admin', desc: 'Active Admin for back office adminstration' },
+    {name: 'api', desc: 'Provide API for mobile device (Beta)' },
+    {name: 'angular', desc: 'Setup an angular application' },
+    {name: 'jazz_hands', desc: 'Better Rails Console tools' }
+  ]
 
-    if all_in || yes?( "Would you like to install splash page?" )
-      generate "happy_seed:splash"
-      packages << "splash"
-    end
-  end
+  run_graph GRAPH
 
-  if all_in || yes?( "Would you like to install devise?" )
-    generate "happy_seed:devise"
-    packages << "devise"
-
-    if all_in || yes?( "Would you like to install devise_invitable?")
-      generate "happy_seed:devise_invitable"
-      packages << "devise_invitable"
-    end
-
-
-    all_connectors = yes?( "Would you like to install all of the oauth connectors?" )
-
-    if all_connectors || yes?( "Would you like to install facebook connect?" )
-      generate "happy_seed:facebook"
-      packages << "facebook"
-    end
-
-    if all_connectors || yes?( "Would you like to install github?" )
-      generate "happy_seed:github"
-      packages << "github"
-    end
-
-    if all_connectors || yes?( "Would you like to install google authentication?" )
-      generate "happy_seed:googleoauth"
-      packages << "googleoauth"
-    end
-
-    if all_connectors || yes?( "Would you like to install instagram?" )
-      generate "happy_seed:instagram"
-      packages << "instagram"
-    end
-
-    if all_connectors || yes?( "Would you like to install twitter?" )
-      generate "happy_seed:twitter"
-      packages << "twitter"
-    end
-  end
-
-  if all_in || yes?( "Would you like to install active admin?" )
-    generate "happy_seed:admin"
-    packages << "admin"
-  end
-
-  if yes?( "(BETA) Would you like to install angular?" )
-    generate "happy_seed:angular_install"
-    packages << "angular"
-  end
+  run "bundle exec spring binstub --all"
 
   puts "Setting up git"
   git :init
   git add: "."
-  git commit: "-a -m 'Based off of happy_seed: #{packages.join( ', ')} included'"
+  git commit: "-a -m 'Based off of happy_seed: #{@packages.join( ', ')} included'"
 end
+